# refactor: remove unnecessary "use client" directives to reduce client bundle size