--- a/components/theme-provider.tsx
+++ b/components/theme-provider.tsx
@@ -1,4 +1,3 @@
-"use client";
 
 import * as React from "react";
 import { ThemeProvider as NextThemesProvider } from "next-themes";