--- a/components/ui/dialog.tsx
+++ b/components/ui/dialog.tsx
@@ -1,4 +1,3 @@
-"use client";
 
 import * as React from "react";
 import * as DialogPrimitive from "@radix-ui/react-dialog";