--- a/components/ui/label.tsx
+++ b/components/ui/label.tsx
@@ -1,4 +1,3 @@
-
 import * as React from "react";
 import * as LabelPrimitive from "@radix-ui/react-label";