--- a/components/ui/popover.tsx
+++ b/components/ui/popover.tsx
@@ -1,4 +1,3 @@
-
 import * as React from "react";
 import * as PopoverPrimitive from "@radix-ui/react-popover";