--- a/components/ui/sonner.tsx
+++ b/components/ui/sonner.tsx
@@ -1,4 +1,3 @@
-
 import { useTheme } from "next-themes";
 import { Toaster as Sonner, type ToasterProps } from "sonner";