--- a/components/ui/sonner.tsx
+++ b/components/ui/sonner.tsx
@@ -1,9 +1,11 @@
-import { useTheme } from "next-themes";
+"use client";
+
 import { Toaster as Sonner, type ToasterProps } from "sonner";
+import { useTheme } from "next-themes";
 
 const Toaster = (props: ToasterProps) => {
   const { theme = "system" } = useTheme();
   return <Sonner theme={theme as ToasterProps["theme"]} {...props} />;
 };
 
-export { Toaster };
+export default Toaster;