--- a/components/ui/tooltip.tsx
+++ b/components/ui/tooltip.tsx
@@ -1,4 +1,3 @@
-
 import * as React from "react";
 import * as TooltipPrimitive from "@radix-ui/react-tooltip";