### Summary
Removed `"use client"` directives from components and pages that do not require client-side rendering. This reduces the client bundle size, improves maintainability, and makes the server/client boundary clearer.

### Changes
- Audited all files with `"use client"` in `app/` and `components/`
- Kept `"use client"` only where necessary:
  - React hooks are used (`useState`, `useEffect`, etc.)
  - Browser APIs are accessed (`window`, `document`, `localStorage`, etc.)
  - Required by Next.js (e.g., error boundaries)
- Removed it from files safe to render as server components

### Why
- **Performance**: Smaller client-side JavaScript footprint
- **Clarity**: Clearer separation of server vs client components
- **Maintainability**: Easier to reason about rendering requirements

### Impact
- Reduced number of client components
- Leaner production build
- No change in app behavior

### Verification
- Ran `npm run lint` → ✅
- Ran `npm test` → ✅
- Verified `npm run dev` works and all affected pages/components behave the same
- Build works locally after resolving unrelated migration issue

---

Closes #397