@gsharma101 2025-08-15
### Performance impact (quantified)

**Bundle sizes (`next build`)**
| Metric | Before | After | Δ | Δ% |
|---|---:|---:|---:|---:|
| First Load JS (shared) | 101 kB | 101 kB | 0 kB | 0% |
| /dashboard | 174 kB | 173 kB | -1 kB | -0.6% |

**Lighthouse (desktop, prod build, median of 3)**
| Metric | Before | After | Δ |
|---|---:|---:|---:|
| FCP | 0.22 s | 0.23 s | +0.01 s |
| LCP | 0.60 s | 0.61 s | +0.01 s |
| TBT | 0 ms | 0 ms | 0 ms |
| TTI | 0.60 s | 0.61 s | +0.01 s |

**Interpretation:** The change has **negligible bundle and UX impact** (≈0–1 kB; ≤0.01 s). This suggests the refactor is behaviorally safe and does not regress performance.

_Methodology: Same machine/network. `npm run build && npx next start`. Analyzer enabled (`ANALYZE=true`)._

## Before
<img width="1366" height="720" alt="Screenshot 2025-08-16 011233" src="https://github.com/user-attachments/assets/5d4c868f-edf0-4fe8-85d8-2b5aea9d48fe" />

<img width="1366" height="720" alt="image" src="https://github.com/user-attachments/assets/9ed2e256-9f7b-4ff9-9b51-d875711376ee" />

## After
<img width="1366" height="720" alt="image" src="https://github.com/user-attachments/assets/30c175b3-504c-4d42-8aba-8fbb7a1d25f6" />

<img width="1366" height="720" alt="image" src="https://github.com/user-attachments/assets/333ad8ae-4f6d-4c75-a1bf-49382f53235b" />

[build-before.txt](https://github.com/user-attachments/files/21801686/build-before.txt)

[build-after.txt](https://github.com/user-attachments/files/21801688/build-after.txt)


[lh-before-1.json](https://github.com/user-attachments/files/21801703/lh-before-1.json)
[lh-before-2.json](https://github.com/user-attachments/files/21801694/lh-before-2.json)
[lh-before-3.json](https://github.com/user-attachments/files/21801696/lh-before-3.json)



[lh-after-1.json](https://github.com/user-attachments/files/21801710/lh-after-1.json)
[lh-after-2.json](https://github.com/user-attachments/files/21801711/lh-after-2.json)
[lh-after-3.json](https://github.com/user-attachments/files/21801712/lh-after-3.json)